Move Gerbil.Fresh -> Fresh
This really has nothing to do with the gerbil integration.

Also adds a short module-level comment.
